--- src/app/home/page.tsx
@@ -1,10 +1,0 @@
-import { Header, SideNav } from '@components';
-
-export default function Home() {
-  return (
-    <div>
-      <Header />
-      <SideNav />
-    </div>
-  );
-}
